Move EFIBase to pyanaconda.bootloader.efi
Move the efi-related classes to a new file.
