+++ pyanaconda/bootloader/__init__.py
@@ -24,6 +24,7 @@ from pyanaconda.core import util
 from blivet.devicelibs import raid
 
 from pyanaconda.bootloader.base import BootLoaderError, Arguments, BootLoader
+from pyanaconda.bootloader.efi import EFIGRUB1, EFIGRUB, Aarch64EFIGRUB, ArmEFIGRUB, MacEFIGRUB
 from pyanaconda.bootloader.grub import GRUB
 from pyanaconda.bootloader.grub2 import GRUB2, IPSeriesGRUB2
 from pyanaconda.bootloader.image import LinuxBootLoaderImage, TbootLinuxBootLoaderImage
@@ -38,230 +39,6 @@ from pyanaconda.core.configuration.anaconda import conf
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
-
-
-class EFIBase(object):
-
-    def __init__(self):
-        super().__init__()
-        self.efi_dir = None
-
-    @property
-    def _config_dir(self):
-        return "efi/EFI/{}".format(self.efi_dir)
-
-    def efibootmgr(self, *args, **kwargs):
-        if not conf.target.is_hardware:
-            log.info("Skipping efibootmgr for image/directory install.")
-            return ""
-
-        if "noefi" in flags.cmdline:
-            log.info("Skipping efibootmgr for noefi")
-            return ""
-
-        if kwargs.pop("capture", False):
-            exec_func = util.execWithCapture
-        else:
-            exec_func = util.execWithRedirect
-        if "root" not in kwargs:
-            kwargs["root"] = util.getSysroot()
-
-        return exec_func("efibootmgr", list(args), **kwargs)
-
-    @property
-    def efi_dir_as_efifs_dir(self):
-        ret = self._config_dir.replace('efi/', '')
-        return "\\" + ret.replace('/', '\\')
-
-    def _add_single_efi_boot_target(self, partition):
-        boot_disk = partition.disk
-        boot_part_num = str(partition.parted_partition.number)
-
-        rc = self.efibootmgr(
-            "-c", "-w", "-L", productName.split("-")[0],  # pylint: disable=no-member
-            "-d", boot_disk.path, "-p", boot_part_num,
-            "-l", self.efi_dir_as_efifs_dir + self._efi_binary,  # pylint: disable=no-member
-            root=util.getSysroot()
-        )
-        if rc:
-            raise BootLoaderError("failed to set new efi boot target. This is most likely a kernel or firmware bug.")
-
-    def add_efi_boot_target(self):
-        if self.stage1_device.type == "partition":  # pylint: disable=no-member
-            self._add_single_efi_boot_target(self.stage1_device)  # pylint: disable=no-member
-        elif self.stage1_device.type == "mdarray":  # pylint: disable=no-member
-            for parent in self.stage1_device.parents:  # pylint: disable=no-member
-                self._add_single_efi_boot_target(parent)
-
-    def remove_efi_boot_target(self):
-        buf = self.efibootmgr(capture=True)
-        for line in buf.splitlines():
-            try:
-                (slot, _product) = line.split(None, 1)
-            except ValueError:
-                continue
-
-            if _product == productName.split("-")[0]:           # pylint: disable=no-member
-                slot_id = slot[4:8]
-                # slot_id is hex, we can't use .isint and use this regex:
-                if not re.match("^[0-9a-fA-F]+$", slot_id):
-                    log.warning("failed to parse efi boot slot (%s)", slot)
-                    continue
-
-                rc = self.efibootmgr("-b", slot_id, "-B")
-                if rc:
-                    raise BootLoaderError("failed to remove old efi boot entry.  This is most likely a kernel or firmware bug.")
-
-    def update(self):
-        self.install()
-
-    def write(self):
-        """ Write the bootloader configuration and install the bootloader. """
-        if self.skip_bootloader:  # pylint: disable=no-member
-            return
-
-        if self.update_only:  # pylint: disable=no-member
-            self.update()
-            return
-
-        try:
-            os.sync()
-            self.stage2_device.format.sync(root=util.getTargetPhysicalRoot()) # pylint: disable=no-member
-            self.install()
-        finally:
-            self.write_config()  # pylint: disable=no-member
-
-    def check(self):
-        return True
-
-    def install(self, args=None):
-        if not self.keep_boot_order:  # pylint: disable=no-member
-            self.remove_efi_boot_target()
-        self.add_efi_boot_target()
-
-
-class EFIGRUB1(EFIBase, GRUB):
-    packages = ["efibootmgr"]
-    can_dual_boot = False
-
-    # list of strings representing options for boot device types
-    stage2_device_types = ["partition"]
-    stage2_raid_levels = []
-    stage2_raid_member_types = []
-    stage2_raid_metadata = []
-
-    stage2_is_valid_stage1 = False
-    stage2_bootable = False
-
-    _efi_binary = "\\grub.efi"
-
-    def __init__(self):
-        super().__init__()
-        self.efi_dir = 'BOOT'
-
-    #
-    # configuration
-    #
-
-    @property
-    def efi_product_path(self):
-        """ The EFI product path.
-
-            eg: HD(1,800,64000,faacb4ef-e361-455e-bd97-ca33632550c3)
-        """
-        buf = self.efibootmgr("-v", capture=True)
-        matches = re.search(productName + r'\s+(HD\(.+?\))', buf)
-        if matches and matches.groups():
-            return matches.group(1)
-        return ""
-
-    @property
-    def grub_conf_device_line(self):
-        return "device %s %s\n" % (self.grub_device_name(self.stage2_device),
-                                   self.efi_product_path)
-
-
-class EFIGRUB(EFIBase, GRUB2):
-    _packages32 = [ "grub2-efi-ia32", "shim-ia32" ]
-    _packages_common = [ "efibootmgr" ]
-    can_dual_boot = False
-    stage2_is_valid_stage1 = False
-    stage2_bootable = False
-
-    _is_32bit_firmware = False
-
-    def __init__(self):
-        super().__init__()
-        self.efi_dir = 'BOOT'
-        self._packages64 = [ "grub2-efi-x64", "shim-x64" ]
-
-        try:
-            f = open("/sys/firmware/efi/fw_platform_size", "r")
-            value = f.readline().strip()
-        except IOError:
-            log.info("Reading /sys/firmware/efi/fw_platform_size failed, defaulting to 64-bit install.")
-            value = '64'
-        if value == '32':
-            self._is_32bit_firmware = True
-
-    @property
-    def _efi_binary(self):
-        if self._is_32bit_firmware:
-            return "\\shimia32.efi"
-        return "\\shimx64.efi"
-
-    @property
-    def packages(self):
-        if self._is_32bit_firmware:
-            return self._packages32 + self._packages_common + \
-                super().packages
-        return self._packages64 + self._packages_common + \
-            super().packages
-
-class Aarch64EFIGRUB(EFIGRUB):
-    _serial_consoles = ["ttyAMA", "ttyS"]
-    _efi_binary = "\\shimaa64.efi"
-
-    def __init__(self):
-        super().__init__()
-        self._packages64 = ["grub2-efi-aa64", "shim-aa64"]
-
-class ArmEFIGRUB(EFIGRUB):
-    _serial_consoles = ["ttyAMA", "ttyS"]
-    _efi_binary = "\\grubarm.efi"
-
-    def __init__(self):
-        super().__init__()
-        self._packages32 = ["grub2-efi-arm"]
-        self._is_32bit_firmware = True
-
-class MacEFIGRUB(EFIGRUB):
-    def __init__(self):
-        super().__init__()
-        self._packages64.extend(["grub2-tools-efi", "mactel-boot"])
-
-    def mactel_config(self):
-        if os.path.exists(util.getSysroot() + "/usr/libexec/mactel-boot-setup"):
-            rc = util.execInSysroot("/usr/libexec/mactel-boot-setup", [])
-            if rc:
-                log.error("failed to configure Mac boot loader")
-
-    def install(self, args=None):
-        super().install()
-        self.mactel_config()
-
-    def is_valid_stage1_device(self, device, early=False):
-        valid = super().is_valid_stage1_device(device, early)
-
-        # Make sure we don't pick the OSX root partition
-        if valid and getattr(device.format, "name", "") != "Linux HFS+ ESP":
-            valid = False
-
-        if hasattr(device.format, "name"):
-            log.debug("device.format.name is '%s'", device.format.name)
-
-        log.debug("MacEFIGRUB.is_valid_stage1_device(%s) returning %s", device.name, valid)
-        return valid
 
 
 # Inherit abstract methods from BootLoader
